global frame_alloc enum made
