+++ kernel/src/memory/mod.rs
@@ -1,4 +1,5 @@
 pub mod boot_frame_allocator;
+pub mod bitmap_frame_allocator;
 pub mod frame_allocator;
 pub mod heap;
 pub mod paging;
